+++ zoom/_assets/standard_apps/sample/widgets/__init__.py
@@ -7,8 +7,10 @@ import zoom.html as h
 
 from zoom.components.widgets.progress import ProgressWidget
 from zoom.components.widgets.cards import Card
+from zoom.components.widgets.metrics import Metric, MetricWidget
 
 from components.widgets_layout import WidgetsLayout
+
 
 
 class Cards(zoom.DynamicComponent):
@@ -16,6 +18,8 @@ class Cards(zoom.DynamicComponent):
 
 
 def view():
+
+    layout = WidgetsLayout()
 
     card = Card()
     cards = Cards().format(
@@ -39,11 +43,43 @@ def view():
     )
     progress = zoom.Component(
         h.h2('ProgressWidget'),
-        WidgetsLayout().format(*progress_widgets),
+        layout.format(*progress_widgets),
+    )
+
+    metrics = [
+        Metric(
+            title='Pipeline',
+            format='${:,.2f}',
+        ),
+        Metric(
+            title='Inventory',
+            classed='bg-gradient-info'
+        ),
+        Metric(
+            title='Expenses',
+            classed='bg-gradient-warning',
+            data=[100, 200, 700, 400],
+            labels=['January', 'February', 'March', 'April']
+        ),
+        Metric(
+            title='Errors',
+            classed='bg-gradient-danger'
+        ),
+    ]
+
+    metric_widget = MetricWidget()
+    metric_widgets = (
+        card.format(metric_widget.format(metric)) for metric in metrics
+    )
+
+    metrics_section = zoom.Component(
+        h.h2('MetricWidget'),
+        layout.format(*metric_widgets)
     )
 
     return zoom.page(
         cards,
         progress,
+        metrics_section,
         title='Widgets'
     )