+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add partial response
 - add cards
 - add progress widget
+- add metric widget
 
 ### Added